tests/wpt/web-platform-tests/xhr/abort-during-upload.htm

<!doctype html>
<html>
  <head>
    <title>XMLHttpRequest: abort() while sending data</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <script src="resources/xmlhttprequest-event-order.js"></script>
    <link rel="help" href="https://xhr.spec.whatwg.org/#the-abort()-method" data-tested-assertations="following-sibling::ol/li[4]/ol/li[7] following-sibling::ol/li[4]/ol/li[7]/ol/li[2] following-sibling::ol/li[4]/ol/li[7]/ol/li[3] following-sibling::ol/li[4]/ol/li[7]/ol/li[4]" />
    <link rel="help" href="https://xhr.spec.whatwg.org/#make-upload-progress-notifications" data-tested-assertations="following::ul[1]/li[1] following::ul[1]/li[2]/ol[1]/li[2] following::ul[1]/li[2]/ol[1]/li[3] following::ul[1]/li[2]/ol[1]/li[4]" />
  </head>
  <body>
    <div id="log"></div>
    <script>
      var test = async_test(document.title, {timeout:1100})
      test.step(function() {
        var client = new XMLHttpRequest()
        prepare_xhr_for_event_order_test(client);
        client.open("POST", "resources/delay.py?ms=1000")
        client.addEventListener("loadend", function(e) {
          test.step(function() {
            assert_xhr_event_order_matches([1, "loadstart(0,0,false)", "upload.loadstart(0,9999,true)", 4, "upload.abort(0,0,false)", "upload.loadend(0,0,false)", "abort(0,0,false)", "loadend(0,0,false)"]);
            test.done()
          })
        });
        client.send((new Array(10000)).join('a'))
        client.abort()
      })
    </script>
  </body>
</html>
