LayoutTests/imported/w3c/web-platform-tests/xhr/abort-during-upload-expected.txt


FAIL XMLHttpRequest: abort() while sending data assert_equals: expected "upload.loadstart(0,9999,true)" but got "upload.loadstart(0,0,false)"
